Adapt to changes in API for random in Suite (#263)
* wip

* Rework the Proxy and proxy test

Make them work correctly when the ports are chosen by the
operating system, so that tests are more robust w.r.t. other
things running on the system.

Closes #262

* Make go vet happy.

+++ simul/build.go
@@ -179,12 +179,12 @@ func RunTest(rc *platform.RunConfig) (*monitor.Stats, error) {
 		return rs, err
 	}
 
-	monitor.SinkPort = monitorPort
+	monitor.SinkPort = uint16(monitorPort)
 	if err := deployP.Cleanup(); err != nil {
 		log.Error(err)
 		return rs, err
 	}
-	monitor.SinkPort = monitorPort
+	monitor.SinkPort = uint16(monitorPort)
 	done := make(chan error)
 	go func() {
 		done <- monitor.Listen()
